+++ api/helpers/meta/change.js
@@ -144,6 +144,10 @@ module.exports = {
         return moment(value).isValid()
       }
     },
+    queueCalculating: {
+      type: 'boolean',
+      description: 'If true, do not consider queueFinish as accurate until this changes to false.'
+    },
     trackFinish: {
       type: 'string',
       allowNull: true,
@@ -151,6 +155,10 @@ module.exports = {
       custom: function (value) {
         return moment(value).isValid()
       }
+    },
+    showCountdown: {
+      type: 'boolean',
+      description: 'If true, display signs should count down to when the DJ/host goes live.'
     },
     queueMusic: {
       type: 'boolean',